UploadRequest.AllBytesWritten flag fix & BinaryContent prevent external stream dispose
