+++ VimeoDotNet/VimeoDotNet/VimeoClient_Async.cs
@@ -185,603 +185,7 @@ namespace VimeoDotNet
         }
 
         #endregion
-
-        #region Videos
-
-        public async Task<Paginated<Video>> GetVideosAsync()
-        {
-            try
-            {
-                IApiRequest request = GenerateVideosRequest();
-                IRestResponse<Paginated<Video>> response = await request.ExecuteRequestAsync<Paginated<Video>>();
-                CheckStatusCodeError(response, "Error retrieving account videos.");
-
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving account videos.", ex);
-            }
-        }
-
-        public async Task<Video> GetVideoAsync(long clipId)
-        {
-            try
-            {
-                IApiRequest request = GenerateVideosRequest(clipId: clipId);
-                IRestResponse<Video> response = await request.ExecuteRequestAsync<Video>();
-                CheckStatusCodeError(response, "Error retrieving account video.", HttpStatusCode.NotFound);
-
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving account video.", ex);
-            }
-        }
-
-        public async Task<Paginated<Video>> GetUserVideosAsync(long userId)
-        {
-            try
-            {
-                IApiRequest request = GenerateVideosRequest(userId);
-                IRestResponse<Paginated<Video>> response = await request.ExecuteRequestAsync<Paginated<Video>>();
-                CheckStatusCodeError(response, "Error retrieving user videos.", HttpStatusCode.NotFound);
-
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return new Paginated<Video>
-                    {
-                        data = new List<Video>(),
-                        page = 0,
-                        total = 0
-                    };
-                }
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving user videos.", ex);
-            }
-        }
-
-        public async Task<Video> GetUserVideoAsync(long userId, long clipId)
-        {
-            try
-            {
-                IApiRequest request = GenerateVideosRequest(userId, clipId);
-                IRestResponse<Video> response = await request.ExecuteRequestAsync<Video>();
-                CheckStatusCodeError(response, "Error retrieving user video.", HttpStatusCode.NotFound);
-
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving user video.", ex);
-            }
-        }
-
-        public async Task<Paginated<Video>> GetAlbumVideosAsync(long albumId)
-        {
-            try
-            {
-                IApiRequest request = GenerateAlbumVideosRequest(albumId);
-                IRestResponse<Paginated<Video>> response = await request.ExecuteRequestAsync<Paginated<Video>>();
-                CheckStatusCodeError(response, "Error retrieving account album videos.");
-
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving account album videos.", ex);
-            }
-        }
-
-        public async Task<Video> GetAlbumVideoAsync(long albumId, long clipId)
-        {
-            try
-            {
-                IApiRequest request = GenerateAlbumVideosRequest(albumId, clipId: clipId);
-                IRestResponse<Video> response = await request.ExecuteRequestAsync<Video>();
-                CheckStatusCodeError(response, "Error retrieving user album video.", HttpStatusCode.NotFound);
-
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving user album video.", ex);
-            }
-        }
-
-        public async Task<Paginated<Video>> GetUserAlbumVideosAsync(long userId, long albumId)
-        {
-            try
-            {
-                IApiRequest request = GenerateAlbumVideosRequest(albumId, userId);
-                IRestResponse<Paginated<Video>> response = await request.ExecuteRequestAsync<Paginated<Video>>();
-                CheckStatusCodeError(response, "Error retrieving user album videos.", HttpStatusCode.NotFound);
-
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return new Paginated<Video>
-                    {
-                        data = new List<Video>(),
-                        page = 0,
-                        total = 0
-                    };
-                }
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving user album videos.", ex);
-            }
-        }
-
-        public async Task<Video> GetUserAlbumVideoAsync(long userId, long albumId, long clipId)
-        {
-            try
-            {
-                IApiRequest request = GenerateAlbumVideosRequest(albumId, userId, clipId);
-                IRestResponse<Video> response = await request.ExecuteRequestAsync<Video>();
-                CheckStatusCodeError(response, "Error retrieving user album video.", HttpStatusCode.NotFound);
-
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error retrieving user album video.", ex);
-            }
-        }
-
-        public async Task UpdateVideoMetadataAsync(long clipId, VideoUpdateMetadata metaData)
-        {
-            try
-            {
-                IApiRequest request = GenerateVideoPatchRequest(clipId, metaData);
-                IRestResponse response = await request.ExecuteRequestAsync();
-                CheckStatusCodeError(response, "Error updating user video metadata.");
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error updating user video metadata.", ex);
-            }
-        }
-
-        public async Task DeleteVideoAsync(long clipId)
-        {
-            try
-            {
-                IApiRequest request = GenerateVideoDeleteRequest(clipId);
-                IRestResponse response = await request.ExecuteRequestAsync();
-                CheckStatusCodeError(response, "Error deleting video.");
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoApiException("Error updating user video metadata.", ex);
-            }
-        }
-       
-        private IApiRequest GenerateVideosRequest(long? userId = null, long? clipId = null)
-        {
-            ThrowIfUnauthorized();
-
-            IApiRequest request = _apiRequestFactory.GetApiRequest(AccessToken);
-            string endpoint = clipId.HasValue ? Endpoints.UserVideo : Endpoints.UserVideos;
-            request.Method = Method.GET;
-            request.Path = userId.HasValue ? endpoint : Endpoints.GetCurrentUserEndpoint(endpoint);
-
-            if (userId.HasValue)
-            {
-                request.UrlSegments.Add("userId", userId.ToString());
-            }
-            if (clipId.HasValue)
-            {
-                request.UrlSegments.Add("clipId", clipId.ToString());
-            }
-
-            return request;
-        }
-
-        private IApiRequest GenerateAlbumVideosRequest(long albumId, long? userId = null, long? clipId = null)
-        {
-            ThrowIfUnauthorized();
-
-            IApiRequest request = _apiRequestFactory.GetApiRequest(AccessToken);
-            string endpoint = clipId.HasValue ? Endpoints.UserAlbumVideo : Endpoints.UserAlbumVideos;
-            request.Method = Method.GET;
-            request.Path = userId.HasValue ? endpoint : Endpoints.GetCurrentUserEndpoint(endpoint);
-
-            request.UrlSegments.Add("albumId", albumId.ToString());
-            if (userId.HasValue)
-            {
-                request.UrlSegments.Add("userId", userId.ToString());
-            }
-            if (clipId.HasValue)
-            {
-                request.UrlSegments.Add("clipId", clipId.ToString());
-            }
-
-            return request;
-        }
-
-        private IApiRequest GenerateVideoPatchRequest(long clipId, VideoUpdateMetadata metaData)
-        {
-            ThrowIfUnauthorized();
-
-            IApiRequest request = _apiRequestFactory.GetApiRequest(AccessToken);
-            request.Method = Method.PATCH;
-            request.Path = Endpoints.Video;
-
-            request.UrlSegments.Add("clipId", clipId.ToString());
-            if (metaData.Name != null)
-            {
-                request.Query.Add("name", metaData.Name.Trim());
-            }
-            if (metaData.Description != null)
-            {
-                request.Query.Add("description", metaData.Description.Trim());
-            }
-            if (metaData.Privacy != VideoPrivacyEnum.Unknown)
-            {
-                request.Query.Add("privacy.view", metaData.Privacy.ToString().ToLower());
-            }
-            if (metaData.EmbedPrivacy != VideoEmbedPrivacyEnum.Unknown)
-            {
-                request.Query.Add("privacy.embed", metaData.EmbedPrivacy.ToString().ToLower());
-            }
-            request.Query.Add("review_link", metaData.ReviewLinkEnabled.ToString().ToLower());
-
-            return request;
-        }
-
-        private IApiRequest GenerateVideoDeleteRequest(long clipId)
-        {
-            ThrowIfUnauthorized();
-
-            IApiRequest request = _apiRequestFactory.GetApiRequest(AccessToken);
-            request.Method = Method.DELETE;
-            request.Path = Endpoints.Video;
-
-            request.UrlSegments.Add("clipId", clipId.ToString());
-            
-            return request;
-        }
-
-        #endregion
-
-        #region Upload
-
-        public async Task<IUploadRequest> UploadEntireFileAsync(IBinaryContent fileContent,
-            int chunkSize = DEFAULT_UPLOAD_CHUNK_SIZE)
-        {
-            IUploadRequest uploadRequest = await StartUploadFileAsync(fileContent, chunkSize);
-
-            VerifyUploadResponse uploadStatus = null;
-            while (!uploadRequest.IsVerifiedComplete)
-            {
-                uploadStatus = await ContinueUploadFileAsync(uploadRequest);
-                uploadRequest.BytesWritten = uploadStatus.BytesWritten;
-
-                if (uploadRequest.AllBytesWritten)
-                {
-                    // We presumably wrote all the bytes in the file, so verify with Vimeo that it is completed
-                    uploadStatus = await VerifyUploadFileAsync(uploadRequest);
-                    if (uploadStatus.Status == UploadStatusEnum.Completed)
-                    {
-                        // If completed, mark file as complete
-                        await CompleteFileUploadAsync(uploadRequest);
-                        uploadRequest.IsVerifiedComplete = true;
-                    }
-                    else if (uploadStatus.BytesWritten == uploadRequest.FileLength)
-                    {
-                        // Supposedly all bytes are written, but Vimeo doesn't think so, so just bail out
-                        throw new VimeoUploadException(
-                            string.Format(
-                                "Vimeo failed to mark file as completed, Bytes Written: {0:N0}, Expected: {1:N0}.",
-                                uploadStatus.BytesWritten),
-                            uploadRequest);
-                    }
-                }
-            }
-
-            return uploadRequest;
-        }
-
-        public async Task<IUploadRequest> StartUploadFileAsync(IBinaryContent fileContent,
-            int chunkSize = DEFAULT_UPLOAD_CHUNK_SIZE)
-        {
-            if (!fileContent.Data.CanRead)
-            {
-                throw new ArgumentException("fileContent should be readable");
-            }
-            if (fileContent.Data.CanSeek && fileContent.Data.Position > 0)
-            {
-                fileContent.Data.Position = 0;
-            }
-
-            UploadTicket ticket = await GetUploadTicketAsync();
-
-            var uploadRequest = new UploadRequest
-            {
-                Ticket = ticket,
-                File = fileContent,
-                ChunkSize = chunkSize
-            };
-
-            VerifyUploadResponse uploadStatus = await ContinueUploadFileAsync(uploadRequest);
-            uploadRequest.BytesWritten = uploadStatus.BytesWritten;
-
-            return uploadRequest;
-        }
-
-        public async Task<UploadTicket> GetUploadTicketAsync()
-        {
-            try
-            {
-                IApiRequest request = GenerateUploadTicketRequest();
-                IRestResponse<UploadTicket> response = await request.ExecuteRequestAsync<UploadTicket>();
-                CheckStatusCodeError(null, response, "Error generating upload ticket.");
-
-                return response.Data;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoUploadException("Error generating upload ticket.", null, ex);
-            }
-        }
-
-        public async Task<VerifyUploadResponse> ContinueUploadFileAsync(IUploadRequest uploadRequest)
-        {
-            if (uploadRequest.AllBytesWritten)
-            {
-                // Already done, there's nothing to do.
-                return new VerifyUploadResponse
-                {
-                    Status = UploadStatusEnum.InProgress,
-                    BytesWritten = uploadRequest.BytesWritten
-                };
-            }
-
-            try
-            {
-                IApiRequest request = await GenerateFileStreamRequest(uploadRequest.File, uploadRequest.Ticket,
-                    chunkSize: uploadRequest.ChunkSize, written: uploadRequest.BytesWritten);
-                IRestResponse response = await request.ExecuteRequestAsync();
-                CheckStatusCodeError(uploadRequest, response, "Error uploading file chunk.", HttpStatusCode.BadRequest);
-
-                if (response.StatusCode == HttpStatusCode.BadRequest)
-                {
-                    // something went wrong, figure out where we need to start over
-                    return await VerifyUploadFileAsync(uploadRequest);
-                }
-
-                // Success, update total written
-                uploadRequest.BytesWritten += uploadRequest.ChunkSize;
-                uploadRequest.BytesWritten = Math.Min(uploadRequest.BytesWritten, uploadRequest.FileLength);
-                return new VerifyUploadResponse
-                {
-                    Status = UploadStatusEnum.InProgress,
-                    BytesWritten = uploadRequest.BytesWritten
-                };
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoUploadException("Error uploading file chunk", uploadRequest, ex);
-            }
-        }
-
-        public async Task<VerifyUploadResponse> VerifyUploadFileAsync(IUploadRequest uploadRequest)
-        {
-            try
-            {
-                IApiRequest request =
-                    await GenerateFileStreamRequest(uploadRequest.File, uploadRequest.Ticket, verifyOnly: true);
-                IRestResponse response = await request.ExecuteRequestAsync();
-                var verify = new VerifyUploadResponse();
-                CheckStatusCodeError(uploadRequest, response, "Error verifying file upload.", (HttpStatusCode)308);
-
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    verify.BytesWritten = uploadRequest.FileLength;
-                    verify.Status = UploadStatusEnum.Completed;
-                }
-                else if (response.StatusCode == (HttpStatusCode)308)
-                {
-                    verify.Status = UploadStatusEnum.InProgress;
-                    int startIndex = 0;
-                    int endIndex = 0;
-                    Parameter rangeHeader =
-                        response.Headers.FirstOrDefault(h => string.Compare(h.Name, "Range", true) == 0);
-                    if (rangeHeader != null && rangeHeader.Value != null)
-                    {
-                        Match match = _rangeRegex.Match(rangeHeader.Value as string);
-                        if (match.Success
-                            && int.TryParse(match.Groups["start"].Value, out startIndex)
-                            && int.TryParse(match.Groups["end"].Value, out endIndex))
-                        {
-                            verify.BytesWritten = endIndex - startIndex;
-                            if (verify.BytesWritten == uploadRequest.FileLength)
-                            {
-                                verify.Status = UploadStatusEnum.Completed;
-                            }
-                        }
-                    }
-                }
-                else
-                {
-                    verify.Status = UploadStatusEnum.NotFound;
-                }
-
-                return verify;
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoUploadException("Error verifying file upload.", uploadRequest, ex);
-            }
-        }
-
-        public async Task CompleteFileUploadAsync(IUploadRequest uploadRequest)
-        {
-            ThrowIfUnauthorized();
-
-            try
-            {
-                IApiRequest request = GenerateCompleteUploadRequest(uploadRequest.Ticket);
-                IRestResponse response = await request.ExecuteRequestAsync();
-                CheckStatusCodeError(uploadRequest, response, "Error marking file upload as complete.");
-
-                Parameter locationHeader =
-                    response.Headers.FirstOrDefault(h => string.Compare(h.Name, "Location", true) == 0);
-                if (locationHeader != null && locationHeader.Value != null)
-                {
-                    uploadRequest.ClipUri = locationHeader.Value as string;
-                }
-            }
-            catch (Exception ex)
-            {
-                if (ex is VimeoApiException)
-                {
-                    throw;
-                }
-                throw new VimeoUploadException("Error marking file upload as complete.", uploadRequest, ex);
-            }
-        }
-
-        private IApiRequest GenerateUploadTicketRequest()
-        {
-            ThrowIfUnauthorized();
-
-            IApiRequest request = _apiRequestFactory.GetApiRequest(AccessToken);
-            request.Method = Method.POST;
-            request.Path = Endpoints.UploadTicket;
-            request.Query.Add("type", "streaming");
-            return request;
-        }
-
-        private async Task<IApiRequest> GenerateFileStreamRequest(IBinaryContent fileContent, UploadTicket ticket,
-            long written = 0, int? chunkSize = null, bool verifyOnly = false)
-        {
-            if (ticket == null || string.IsNullOrWhiteSpace(ticket.ticket_id))
-            {
-                throw new ArgumentException("Invalid upload ticket.");
-            }
-            if (fileContent.Data.Length > ticket.user.upload_quota.space.free)
-            {
-                throw new InvalidOperationException(
-                    "User does not have enough free space to upload this video. Remaining space: " +
-                    ticket.quota.free_space + ".");
-            }
-
-            IApiRequest request = _apiRequestFactory.GetApiRequest();
-            request.Method = Method.PUT;
-            request.ExcludeAuthorizationHeader = true;
-            request.Path = ticket.upload_link_secure;
-            request.Headers.Add(Request.HeaderContentType, fileContent.ContentType);
-            if (verifyOnly)
-            {
-                request.Headers.Add(Request.HeaderContentLength, "0");
-                request.Headers.Add(Request.HeaderContentRange, "bytes */*");
-            }
-            else
-            {
-                if (chunkSize.HasValue)
-                {
-                    long startIndex = fileContent.Data.CanSeek ? fileContent.Data.Position : written;
-                    long endIndex = Math.Min(startIndex + chunkSize.Value, fileContent.Data.Length);
-                    request.Headers.Add(Request.HeaderContentLength, (endIndex - startIndex).ToString());
-                    request.Headers.Add(Request.HeaderContentRange,
-                        string.Format("bytes {0}-{1}/{2}", startIndex, endIndex, fileContent.Data.Length));
-                    request.BinaryContent = await fileContent.ReadAsync(startIndex, endIndex);
-                }
-                else
-                {
-                    request.Headers.Add(Request.HeaderContentLength, fileContent.Data.Length.ToString());
-                    request.BinaryContent = await fileContent.ReadAllAsync();
-                }
-            }
-
-            return request;
-        }
-
-        private IApiRequest GenerateCompleteUploadRequest(UploadTicket ticket)
-        {
-            IApiRequest request = _apiRequestFactory.GetApiRequest(AccessToken);
-            request.Method = Method.DELETE;
-            request.Path = ticket.complete_uri;
-            return request;
-        }
-
-        #endregion
-
+        
         #region Helper Functions
 
         private void ThrowIfUnauthorized()
