+++ VimeoDotNet/VimeoDotNet/VimeoDotNet.csproj
@@ -125,6 +125,8 @@
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
+    <Compile Include="VimeoClient_Videos.cs" />
+    <Compile Include="VimeoClient_Upload.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
